Fixes a build break caused by r10178. r10178 compares an int value 'written' with a size_t value 'sizeof(num_open_files)' in "base/process_util_unittest.cc" and it caused a compilation error on Mac and Linux. As a quick fix, this change casts the int value 'written' to size_t.
TBR=xji
Review URL: http://codereview.chromium.org/28019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@10179 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/process_util_unittest.cc
@@ -182,7 +182,7 @@ MULTIPROCESS_TEST_MAIN(ProcessUtilsLeakFDChildProcess) {
   num_open_files -= expected_num_open_fds;
 
   int written = write(write_pipe, &num_open_files, sizeof(num_open_files));
-  DCHECK_EQ(written, sizeof(num_open_files));
+  DCHECK_EQ(static_cast<size_t>(written), sizeof(num_open_files));
   close(write_pipe);
 
   return 0;
